view field index to float

+++ server/src/database/migrations/20230727124244_add_view_fields_table/migration.sql
@@ -2,7 +2,7 @@
 CREATE TABLE "viewFields" (
     "id" TEXT NOT NULL,
     "fieldName" TEXT NOT NULL,
-    "index" INTEGER NOT NULL,
+    "index" DOUBLE PRECISION NOT NULL,
     "isVisible" BOOLEAN NOT NULL,
     "objectName" TEXT NOT NULL,
     "sizeInPx" INTEGER NOT NULL,
